Refactor, the api/views/signals.py was becomming a large file. Splitted it up into several small files. Also moved it to a folder and moved the signal_context.py and the pdf.py to the newly created folders. Fixed imports after split and move.
