+++ api/app/signals/apps/api/views/__init__.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 from signals.apps.api.views.area import PrivateAreasViewSet, PublicAreasViewSet
 from signals.apps.api.views.attachment import (
     PrivateSignalAttachmentsViewSet,
@@ -12,15 +12,15 @@ from signals.apps.api.views.departments import PrivateDepartmentViewSet
 from signals.apps.api.views.expression import PrivateExpressionViewSet
 from signals.apps.api.views.ml_tool_proxy import LegacyMlPredictCategoryView
 from signals.apps.api.views.namespace import NamespaceView
-from signals.apps.api.views.pdf import GeneratePdfView
 from signals.apps.api.views.questions import PublicQuestionViewSet
-from signals.apps.api.views.signal import (
-    PrivateSignalViewSet,
-    PublicSignalMapViewSet,
-    PublicSignalViewSet,
+from signals.apps.api.views.signals.private.pdf import GeneratePdfView
+from signals.apps.api.views.signals.private.signal_context import SignalContextViewSet
+from signals.apps.api.views.signals.private.signals import PrivateSignalViewSet
+from signals.apps.api.views.signals.private.signals_promoted_to_parent import (
     SignalPromotedToParentViewSet
 )
-from signals.apps.api.views.signal_context import SignalContextViewSet
+from signals.apps.api.views.signals.public.signal import PublicSignalViewSet
+from signals.apps.api.views.signals.public.signals_map import PublicSignalMapViewSet
 from signals.apps.api.views.source import PrivateSourcesViewSet
 from signals.apps.api.views.status_message_template import StatusMessageTemplatesViewSet
 from signals.apps.api.views.stored_signal_filter import StoredSignalFilterViewSet
