+++ api/app/signals/apps/api/views/signals/private/signal_context.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2021 Gemeente Amsterdam
+# Copyright (C) 2021 - 2022 Gemeente Amsterdam
 import logging
 
 from datapunt_api.rest import HALPagination
