+++ api/app/signals/apps/api/views/signals/private/signals.py
@@ -1,47 +1,36 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 import logging
 
 from datapunt_api.rest import DatapuntViewSet, HALPagination
 from django.conf import settings
 from django.core.exceptions import ValidationError
-from django.db import connection
 from django.db.models import CharField, Value
 from django.db.models.functions import JSONObject
-from django.http import Http404
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import status
 from rest_framework.decorators import action
 from rest_framework.exceptions import NotFound, PermissionDenied
 from rest_framework.exceptions import ValidationError as DRFValidationError
-from rest_framework.generics import get_object_or_404
-from rest_framework.renderers import BrowsableAPIRenderer
 from rest_framework.response import Response
-from rest_framework.viewsets import GenericViewSet, ViewSet
 
 from signals.apps.api.app_settings import SIGNALS_API_GEO_PAGINATE_BY
-from signals.apps.api.filters import SignalFilterSet, SignalPromotedToParentFilter
+from signals.apps.api.filters import SignalFilterSet
 from signals.apps.api.generics import mixins
 from signals.apps.api.generics.filters import FieldMappingOrderingFilter
 from signals.apps.api.generics.pagination import LinkHeaderPaginationForQuerysets
 from signals.apps.api.generics.permissions import (
-    SIAPermissions,
     SignalCreateInitialPermission,
     SignalViewObjectPermission
 )
-from signals.apps.api.renderers import SerializedJsonRenderer
 from signals.apps.api.serializers import (
     AbridgedChildSignalSerializer,
     HistoryHalSerializer,
     PrivateSignalSerializerDetail,
-    PrivateSignalSerializerList,
-    PublicSignalCreateSerializer,
-    PublicSignalSerializerDetail,
-    SignalIdListSerializer
+    PrivateSignalSerializerList
 )
 from signals.apps.api.serializers.email_preview import EmailPreviewSerializer
 from signals.apps.api.serializers.signal_history import HistoryLogHalSerializer
-from signals.apps.api.views._base import PublicSignalGenericViewSet
 from signals.apps.email_integrations.utils import trigger_mail_action_for_email_preview
 from signals.apps.history.models import Log
 from signals.apps.signals import workflow
@@ -49,87 +38,8 @@ from signals.apps.signals.models import Signal
 from signals.apps.signals.models.aggregates.json_agg import JSONAgg
 from signals.apps.signals.models.functions.asgeojson import AsGeoJSON
 from signals.auth.backend import JWTAuthBackend
-from signals.throttling import PostOnlyNoUserRateThrottle
 
 logger = logging.getLogger(__name__)
-
-
-class PublicSignalViewSet(PublicSignalGenericViewSet):
-    throttle_classes = (PostOnlyNoUserRateThrottle,)
-    serializer_class = PublicSignalSerializerDetail
-
-    def list(self, *args, **kwargs):
-        raise Http404
-
-    def create(self, request):
-        serializer = PublicSignalCreateSerializer(data=request.data, context=self.get_serializer_context())
-        serializer.is_valid(raise_exception=True)
-        signal = serializer.save()
-
-        data = PublicSignalSerializerDetail(signal, context=self.get_serializer_context()).data
-        return Response(data, status=status.HTTP_201_CREATED)
-
-    def retrieve(self, request, uuid):
-        signal = get_object_or_404(Signal.objects.all(), uuid=uuid)
-        data = PublicSignalSerializerDetail(signal, context=self.get_serializer_context()).data
-        return Response(data)
-
-
-class PublicSignalMapViewSet(ViewSet):
-    renderer_classes = [SerializedJsonRenderer, BrowsableAPIRenderer]
-    # django-drf has too much overhead with these kinds of 'fast' request.
-    # When implemented using django-drf, retrieving a large number of elements cost around 4s (profiled)
-    # Using pgsql ability to generate geojson, the request time reduces to 30ms (> 130x speedup!)
-    # The downside is that this query has to be (potentially) maintained when changing one of the
-    # following models: signal, categoryassignment, category, location, status
-
-    def list(self, *args, **kwargs):
-        fast_query = f"""
-        select jsonb_build_object(
-            'type', 'FeatureCollection',
-            'features', json_agg(features.feature)
-        ) as result from (
-            select json_build_object(
-                'type', 'Feature',
-                'geometry', st_asgeojson(l.geometrie)::jsonb,
-                'properties', json_build_object(
-                    'id', to_jsonb(s.id),
-                    'created_at', to_jsonb(s.created_at),
-                    'status', to_jsonb(status.state),
-                    'category', json_build_object(
-                        'sub', to_jsonb(cat.name),
-                        'main', to_jsonb(maincat.name)
-                    )
-                )
-            ) as feature
-            from
-                signals_signal s
-                left join signals_categoryassignment ca on s.category_assignment_id = ca.id
-                left join signals_category cat on ca.category_id = cat.id
-                left join signals_category maincat on cat.parent_id = maincat.id,
-                signals_location l,
-                signals_status status
-            where
-                s.location_id = l.id
-                and s.status_id = status.id
-                and status.state not in ('{workflow.AFGEHANDELD}', '{workflow.AFGEHANDELD_EXTERN}', '{workflow.GEANNULEERD}', '{workflow.VERZOEK_TOT_HEROPENEN}')
-            order by s.id desc
-            limit 4000 offset 0
-        ) as features
-        """ # noqa
-
-        cursor = connection.cursor()
-        try:
-            cursor.execute(fast_query)
-            row = cursor.fetchone()
-            return Response(row[0])
-        except Exception as e:
-            cursor.close
-            logger.error('failed to retrieve signals json from db', exc_info=e)
-
-    def get_view_name(self):
-        # Overridden to avoid: "Public Signal Map List" that is the default behavior here.
-        return 'Public Signal Map'
 
 
 class PrivateSignalViewSet(mixins.CreateModelMixin, mixins.UpdateModelMixin, DatapuntViewSet):
@@ -379,16 +289,3 @@ class PrivateSignalViewSet(mixins.CreateModelMixin, mixins.UpdateModelMixin, Dat
 
         headers = self.get_success_headers(serializer.data)
         return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
-
-
-class SignalPromotedToParentViewSet(GenericViewSet, mixins.ListModelMixin):
-    serializer_class = SignalIdListSerializer
-    pagination_class = HALPagination
-
-    authentication_classes = (JWTAuthBackend,)
-    permission_classes = (SIAPermissions,)
-
-    filter_backends = (DjangoFilterBackend,)
-    filterset_class = SignalPromotedToParentFilter
-
-    queryset = Signal.objects.prefetch_related('children').only('id').filter(children__isnull=False).order_by('-id')
